feat(composition): establish runtime wiring
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src/sciretriever/infrastructure/sources/assets/adapters.py
@@ -84,7 +84,7 @@ class AssetFetcherAdapter:
         self,
         transport: BoundedTransportPort,
         *,
-        timeout_seconds: int,
+        timeout_seconds: float,
         max_response_bytes: int,
     ) -> None:
         self._transport = transport
